Extract I/O tap into a module
