+++ lib/Makefile.am
@@ -135,12 +135,13 @@ libtlog_la_SOURCES = \
     source.c                    \
     syslog_json_writer.c        \
     syslog_misc.c               \
+    tap.c                       \
     timespec.c                  \
     tty_sink.c                  \
     tty_source.c                \
     utf8.c
 
-libtlog_la_LIBADD = $(JSON_LIBS) $(LIBCURL) -lrt
+libtlog_la_LIBADD = $(JSON_LIBS) $(LIBCURL) -lutil -lrt
 
 libtlog_test_la_SOURCES = \
     test_json_sink.c        \
